Refactor Compiler, and improve type narrowing

+++ lib/orb/custom_type.ex
@@ -1,5 +1,5 @@
 defmodule Orb.CustomType do
-  @callback wasm_type() :: :i32 | :i64 | :f32 | :f64
+  @callback wasm_type() :: :i64 | :i32 | :f64 | :f32
   @callback load_instruction() :: :load | :load8_s | :load8_u | :load16_s | :load16_u | :load32_s | :load32_u
   # TODO: remove byte_count
   @callback byte_count() :: 1 | 4 | 8
